Use an image instead of canvas for rendering
This fixes the blurry image issue in Safari.  Safari doesn't want to
properly apply image-rendering css to canvas elements, but it handles
img element fine.  This is not an efficient solution, but it'll work for
internal testing.  Before launch, we should definitely:

1. Only apply this hacky fix to Safari users
2. Only refresh the display img element at an interval, using
requestAnimationFrame and checking if the canvas is actually dirty.