Add support for hide_palette query param
This allows hiding several UI components based on the presence of the
hide_palette query param.  The components currently hidden are:

- color palette
- molly-guard
- mute button
- zoom button

This is primarily to support the embeded view in the native apps
listing.  The timer UI is left visible as it seems useful to show to
users without requiring them to load the full webview.