Skip initialization if canvas element isn't found on the page
The js is injected everywhere in the place subreddit, but only the
listing page has the canvas element. If we let the init script run, it
ends up throwing errors when it tries to interact with DOM elements that
don't exist on the page, which might cause other things to break.  We
can fix this by bailing out early if the root element of the place UI
isn't found.