Move canvas drawing into the animationFrame loop
Right now, every time an update is received we update the display.  Due
to the hack we have in place to fix blurriness on Safari, updating the
display is very slow, so any large bursts of updates will likely lock up
the client for a while.  This allows updates to be batched so that the
expensive draw operation is only done at most once every ~16ms.